+++ src/test/scala/faunadb/specs/FileUtils.scala
@@ -4,7 +4,7 @@ import java.io.File
 import java.nio.file.Files
 
 trait FileUtils {
-  def withTempFile[A](name: String)(fn: (File) => A): A = {
+  def withTempFile[A](name: String)(fn: File => A): A = {
     val file = Files.createTempFile(name, "tmp").toFile
     try fn(file) finally file.delete()
   }